Refactor flatten_sagittal into a newly created flattening module. It is needed elsewhere by the API (detect_c2c3) thus it should be in the API.
